printing 0 in task 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long long int i = 1, num = 0, o = 0;
+	unsigned long int i = 1, num = 0, o = 0;
 
 	if (n == 0)
 		_putchar('0');
